backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
websockets==12.0
supabase==2.0.2
pydantic-settings==2.0.3
aiohttp==3.9.1
pydantic[email]==2.5.0
email-validator==2.1.0
